requirements.txt
Django==4.0.6
djangorestframework==3.13.1
Markdown==3.3.7
django-filter==22.1
kavenegar==1.1.2
django-dotenv==1.4.2
kavenegar==1.1.2
drf-yasg==1.21.3